Proofing fixes for “Hans Pfaall”
First up, the file was misnamed, hence content.opf and toc updates. Inside the file, the version that GB had transcribed wasn’t the same as the version in the source scans. I’ve decided (editorial decision) that the source scan version is better, and so have adjusted the transcription to match that.